+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.11.0] - Unreleased
 ### Added
-- `Rabbit.config.receiving_job_class_callable` now receive the full message context (with `message`, `delivery_info` and `arguments` (see the `Rabbit::Receiving::Receive`));
+- `Rabbit.config.receiving_job_class_callable` now receives the full message context (with `message`, `delivery_info` and `arguments` (see the `Rabbit::Receiving::Receive`));
 
 ## [0.10.0] - 2021-03-05
 ### Added